[refactor] deleted_at 유저 엔티티 안으로 수정
